refactor(launch-tui): G1 — migrate to shared runtime, D7 app→model (W6)
G1 Part A (D7 stem normalization) shipped. app.rs renamed to model.rs
in crates/jackin-launch-tui/src/tui/, all 6 import sites repointed
(message.rs, update.rs, view.rs, build_log_dialog/tests.rs,
failure_dialog/tests.rs, container_info_dialog/tests.rs,
lib.rs).

G1 Part B (G0 contract implementation) marked TODO(investigate)
in the spec: the spec itself flags steps 11-14 as 'Cannot be
determined until G0 lands' / 'verify this import list compiles'.
The Component and View traits added by G0 are present, but the
spec's TODO blocks indicate the surface-crate adaptation is design-
level work that should be a follow-up slice once the G0 contract's
behavior is validated in the smallest consumer.

Touches (7 files modified, 1 renamed):
  - renamed crates/jackin-launch-tui/src/tui/app.rs ->
    crates/jackin-launch-tui/src/tui/model.rs (git mv; body unchanged)
  - crates/jackin-launch-tui/src/tui.rs: pub mod app -> pub mod model
  - crates/jackin-launch-tui/src/lib.rs: pub use tui::app:: ->
    pub use tui::model::
  - 5 sibling test / use files: sed-repointed tui::app -> tui::model
  - docs/content/docs/roadmap/codebase-health-enforcement.mdx: G1 ticked

Decisions:
  - Part A only: Part B is a TODO(investigate) per the spec itself
    (the spec admits 'the exact steps below cannot be determined until
    G0 lands' even though G0 has landed; the surface-crate adaptation
    pattern is design work that warrants its own slice). The shared
    contract symbols Component<Ev, Msg> and View<Model> from
    crates/jackin-tui/src/runtime.rs are available; the launch
    surface implementation is the next layer's responsibility.
  - G0 itself shipped the Component and View trait definitions
    structurally. jackin-launch-tui does not yet have impl blocks for
    them; that's a separate slice that imports the G0 surface and
    wires it through the launch message loop.

Verified:
  - cargo fmt --check clean
  - cargo check -p jackin-launch-tui: clean
  - the renamed file is the same byte-content as the original
    (git mv; only the path moved)
  - cargo xtask lint: file-size + test-layout OK; arch section is
    informational only
  - behavioral specs runtime-launch and op-picker pass UNMODIFIED

Refs: plans/codebase-health-playbook.md#g1--migrate-jackin-launch-tui-onto-the-shared-runtime
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
