+++ crates/jackin-launch-tui/src/tui/components/container_info_dialog/tests.rs
@@ -1,7 +1,7 @@
 use ratatui::layout::Rect;
 use ratatui::{Terminal, backend::TestBackend, buffer::Buffer};
 
-use crate::tui::app::{LaunchIdentity, LaunchTargetKind, LaunchView};
+use crate::tui::model::{LaunchIdentity, LaunchTargetKind, LaunchView};
 use crate::tui::update::initial_view;
 use crate::tui::view::render_launch_frame;
