+++ docs/content/docs/roadmap/codebase-health-enforcement.mdx
@@ -261,7 +261,7 @@ Used by every C/E slice. Pure move; contents are not rewritten.
 ### Phase G — unify the TUI runtime (W6; largest)
 
 - [x] **G0 — design + build the shared contract in `jackin-tui` (D5).** `Model`/`Message`/`update` + a `Component`/`View` contract; one central `update` per app. No stack migrated yet — just the runtime + a doc'd contract.
-- [ ] **G1 — migrate `jackin-launch-tui`** onto it (smallest stack; proves the contract).
+- [x] **G1 — migrate `jackin-launch-tui`** onto it (smallest stack; proves the contract).
 - [ ] **G2 — migrate the capsule TUI** onto it.
 - [ ] **G3 — migrate `jackin-console`**, landing [unify-settings-editor-surfaces](/roadmap/unify-settings-editor-surfaces/) on the shared contract so the editor + settings models become one implementation.
